src/backend/requirements.txt
flask
flask_cors
requests
pydub
pocketsphinx
speechrecognition
deepspeech-gpu==0.9.3
GPUtil
tqdm